arcadia-backend/app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
fastapi==0.115.3
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.1.2
packaging==24.1
pdf2image==1.17.0
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pytesseract==0.3.13
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
starlette==0.41.2
threadpoolctl==3.5.0
typing_extensions==4.12.2
uvicorn==0.32.0
